requirements.txt
fastapi
uvicorn
jinja2
python-multipart
aiohttp
paho-mqtt
pyyaml
async-timeout